Corrige 10 bugs encontrados no fluxo inicial de configuração
Plugin QGIS:
- Corrige chave de settings inconsistente (server_url → saved_server) em api_client.py
- Corrige filtro de busca que destruía estado dos checkboxes em manage_users_dialog
- Corrige filtro de busca que destruía seleções em import_users_dialog
- Corrige exibição de data_fim null na tabela de lotes (manage_lotes_dialog)

Server:
- Corrige negação dupla "não não" na mensagem de erro em database_version.js
- Remove parâmetro usuarioUuid não utilizado de deleteLotes em projeto_ctrl.js

Build/Config:
- Adiciona verificação de existência do diretório client/ em create_build.js

Database:
- Corrige tipo de volume_armazenamento_id de SMALLINT para INTEGER em acervo.sql
- Corrige tipo de projeto_id de INTEGER para BIGINT em acervo.sql

Documentação:
- Corrige versão do Express de 4 para 5 em CLAUDE.md

+++ er/acervo.sql
@@ -15,7 +15,7 @@ CREATE TABLE acervo.volume_armazenamento(
 CREATE TABLE acervo.volume_tipo_produto(
 	id SERIAL NOT NULL PRIMARY KEY,
 	tipo_produto_id SMALLINT NOT NULL REFERENCES dominio.tipo_produto (code),
-	volume_armazenamento_id SMALLINT NOT NULL REFERENCES acervo.volume_armazenamento (id),
+	volume_armazenamento_id INTEGER NOT NULL REFERENCES acervo.volume_armazenamento (id),
 	primario BOOLEAN NOT NULL DEFAULT TRUE
 );
 
@@ -36,7 +36,7 @@ CREATE TABLE acervo.projeto (
 
 CREATE TABLE acervo.lote (
     id BIGSERIAL NOT NULL PRIMARY KEY,
-	projeto_id INTEGER NOT NULL REFERENCES acervo.projeto (id),
+	projeto_id BIGINT NOT NULL REFERENCES acervo.projeto (id),
 	pit VARCHAR(255) NOT NULL,
     nome VARCHAR(255) NOT NULL,
     descricao TEXT,
